+++ libft/ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: livsauze <livsauze@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 10:56:49 by livsauze          #+#    #+#             */
-/*   Updated: 2024/11/05 13:58:06 by livsauze         ###   ########.fr       */
+/*   Updated: 2025/01/23 16:45:12 by livsauze         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
 	char	*temp;
-	size_t	i;
+	int	i;
 
 	i = 0;
 	temp = s;
