+++ libft/ft_strlcpy.c
@@ -6,16 +6,16 @@
 /*   By: livsauze <livsauze@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 12:47:59 by livsauze          #+#    #+#             */
-/*   Updated: 2024/11/05 13:59:03 by livsauze         ###   ########.fr       */
+/*   Updated: 2025/01/23 16:45:50 by livsauze         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcpy(char *dest, const char *src, size_t size)
+int	ft_strlcpy(char *dest, const char *src, int size)
 {
-	size_t	i;
-	size_t	lens;
+	int	i;
+	int	lens;
 
 	lens = ft_strlen(src);
 	i = 0;
